requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
attrs==25.4.0
boolean.py==5.0
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
cicflowmeter==0.5.0
click==8.3.1
cyclonedx-bom==7.2.1
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
fastapi==0.129.0
httpx==0.28.1
fqdn==1.5.1
h11==0.16.0
httptools==0.7.1
idna==3.11
isoduration==20.11.0
joblib==1.5.3
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
license-expression==30.4.4
lxml==6.0.2
numpy==2.4.2
packageurl-python==0.17.6
packaging==25.0
pandas==3.0.0
pip-requirements-parser==32.0.1
py-serializable==2.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scapy==2.7.0
psutil==6.1.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sortedcontainers==2.4.0
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.41.0
# uvloop: Unix/macOS only; omit on Windows. Uvicorn uses asyncio on Windows.
watchfiles==1.1.1
webcolors==25.10.0
websockets==16.0
xgboost==3.2.0
pytest==8.4.2
redis==5.2.1
